feat: Simplify Angstrom indexing (#323)
* feat: Simplify Angstrom indexing

As it was before, we would be emitting EntityChanges every time there would be a swap on an Angstrom pool, not necessarily when the Angstrom Config changed.
To simplify this, we need to get rid of the store for the AngstromConfig entirely and just emit it when it actually happens!

#time 1h 8m


#time 2m

* chore: Use substreams advanced params for Angstrom

#time 15m

* fix(angstrom): Add attributes to EntityChanges if it exists, create new one if not

#time 36m

* chore(angstrom): Remove debug logs and rename private fn

#time 7m